A separate #resend_email_change for /account
Until now, Users#resend_email_change was used in two distinct scenarios:

1. Admins performing the action for other users' accounts
2. Users (of all roles) performing the action for their own account

It seems as though the `if @user.normal?` conditional was actually
breaking this distinction: in the case of admin users in scenario 2,
they'd only ever see the flash notice intended for scenario 1.

Scenario 2 is now exclusively performed from the "Change your email or
password" page. That page has already moved to the /account namespace
because it's becoming a sub-page of the new Account page. That page now
has its own dedicated `#resend_email_change` endpoint.

Scenario 1 is still performed from Users#edit.

Similarly to the approach with the actions that I've just
lifted-and-shifted from UsersController to
Account::EmailPasswordsController, for the time being I'm keeping the
naming/RESTful design as I found it and planning to rethink that after
the upcoming design review.