A separate #cancel_email_change for /account
Until now, Users#cancel_email_change was used in two distinct scenarios:

1. Admins performing the action for other users' accounts
2. Users (of all roles) performing the action for their own account

Scenario 2 is now exclusively performed from the "Change your email or
password" page. That page has already moved to the /account namespace
because it's becoming a sub-page of the new Account page.

(Scenario 1 is still performed from Users#edit.)

Similarly to the approach with the actions that I've just
lifted-and-shifted from UsersController to
Account::EmailPasswordsController, for the time being I'm keeping the
naming/RESTful design as I found it and planning to rethink that after
the upcoming design review.

I'm struggling to fully understand the UserPolicy (!) but I think it was
safe to remove the remaining references to `allow_self_only`, since
Users#cancel_email_change was the last UsersController action left that
was used in the ways described by both of the scenarios listed above ^