Move 2SV setup page to /account
I'm planning to make this page another sub-page of the new Account page.

For consistency with the Account page and the other Account sub-pages,
I've changed this page's URL to live under /account instead of /users.

Since this page is already only ever used by the current_user to
configure their own account and since the controller already lives in
the fairly neutral `Devise` namespace, it's only the URL that needs
changing to make this an "account" page instead of a "user" one.

And since the affected part of the URL isn't factored into its
Rails-generated URL helper method names, no redirects or links need
updating to complete this change, either.