Introduce new Account page with one sub-page
That "sub-page" is the "Change your email or password" page, which is
linked to from a card component.

I'm planning to add 4 more cards to this page in later commits. Like
this first one, they'll link to a standalone sub-page each allowing
users to make changes to a different aspect of their account.

In this commit I've also removed an integration test instead of updating
it to reflect my changes, because the important stuff to cover is now
covered by the AccountTest.