Strip whitespace from name on existing Organisation records
In #2373 I noticed that the query we'd used to do this in #2351 didn't
escape the regular expression properly. ActiveRecord seems to have
generated the following SQL WHERE expression:

    WHERE (name REGEXP('^ +') OR name REGEXP(' +$')

I _think_ this will only have found names with leading/trailing
space characters and not leading/trailing
newlines/carriage-returns/tabs. Thus the previous migration didn't do
anything bad, it just may have missed _some_ records.

In this new version of the migration I've constructed the query slightly
differently. Firstly I've double-escaped the whitespace character in the
regular expression and secondly I've used an Array Condition vs a Pure
String Condition [1] to ensure argument safety. The new query generates
the following SQL WHERE expression:

    WHERE (name REGEXP '^\\s+' OR name REGEXP '\\s+$')

This has the REGEXP string argument escaped correctly.

As per the migrations in #2373 I've use
ActiveRecord::Persistence#update_attribute in order to skip model
validation, in case some of the existing Organisation records are
invalid for other reasons.

As before I did consider writing the migration as a SQL UPDATE, but
generating a TRIM expression that is exactly equivalent to String#strip
isn't as trivial as it sounds!

[1]: https://guides.rubyonrails.org/active_record_querying.html#conditions