Introduce Role and organisation Account sub-page
This provides some functionality that was lost in the switch from
Users#edit to the new Account page.

For Admin and Superadmin users it allows making changes, but for users
with other roles it's read-only.

Since this is another Account sub-page and another mixed bag of a page,
for consistency I've mostly followed the approaches taken by the "Change
your email or password" page.

---

I've let one of the new helper methods be tested through the integration
test only, because I didn't manage to figure out the correct test setup
required for a helper method that's calling a Pundit method.