Namespace the ApplicationPolicy
In order to allow Publishing Managers to remove their own signin
permission from apps I'm going to need an instance of the Application so
that I can check whether it has delegatable permissions.

This preparatory change will allow me to pass an instance of Application
to `authorize` in order to automagically find this
`Account::ApplicationPolicy` class.