Remove out-of-date TODO from InvitationsController
The change described in the TODO seems to have been made in this commit
from 6 years ago [1] which was part of the upgrade to Rails v5.1.4 [2].
However, the TODO comment was not removed at the same time, presumably
due to an oversight.

I've tried to understand more about what the TODO comment meant, but
it's pretty hard to work it out. The Devise::InvitationsController from
the current version of devise_invitable (v2.0.8) inherits from
DeviseController in the current version of devise (v4.9.2). And the
latter implements a `resource_params` method [3] which still uses
`params.fetch()`. However, our InvitationsController which inherits from
Devise::InvitationsController implements its own `resource_params`
method which overwrites (no call to `super`) the implementation in
DeviseController.

The Devise::InvitationsController [4] from devise_invitable doesn't seem
to call the `resource_params` method. However, our InvitationsController
does call our own `resource_params` method.

So things are all a bit confusing - not least because I think some of
the methods in our current InvitationsController seem to be based on an
old implementation of Devise::InvitationsController. However, I don't
propose to address that in this commit.

Anyway, based on the above and the fact that the app seems to have been
working fine for the last 6 years, I think it's safe to delete this TODO
comment! If we don't do it now, we never will!!

[1]: d28ff3c61fb6775a34c08e7a4f8d2dba8121b11e
[2]: #593
[3]: https://github.com/heartcombo/devise/blob/8b0b849a67c46b10827743aa0ccb0679d69e5396/app/controllers/devise_controller.rb#L208-L210
[4]: https://github.com/scambra/devise_invitable/blob/v2.0.8/app/controllers/devise/invitations_controller.rb