Include production environment in email subjects
We've done some user-testing of our email subject lines and users
generally preferred to see the environment name in all cases, even in
the production environment.

I've had to make a small change to the NoisyBatchInvitationTest as it
was also partially testing the subject as well as the email from
address. Changing the email from address is out of scope for this bit
of work.