Remove out-of-date NOTE from InvitationsController
* 1st paragraph of NOTE

> `current_user` doesn't exist for `#edit` and `#update` actions as
> implemented in our current (out-of-date) versions of Devise and
> DeviseInvitable

I'm working on the assumption that "`current_user` doesn't exist" in the
NOTE meant that the return value from the `current_user` method was
`nil` rather than that the `current_user` method did not exist.

The fact that `current_user` is `nil` for `#edit` & `#update` actions is
*still* true for the current version (v4.9.2) of devise and the current
version (v2.0.8) of devise_invitable. There is a slight implication in
the NOTE that this is only true because we're using an out-of-date
version of devise & devise_invitable. However, I'm pretty confident this
is by design in that there is a `require_no_authentication` before
action set for `#edit`, `#update` & `#destroy` [1]. This before action
is defined in `DeviseController` [2].

* 2nd paragraph, 1st sentence of NOTE

> With the old `attr_accessible` approach, this would fall back to the
> default whitelist (i.e. equivalent to the `:normal` role) and this this
> preserves that behaviour.

This seems to be talking about functionality which was removed 8 years
ago [3] as part of the upgrade to Rails v4 [4]. In turn that
functionality seems to have been added [5] just before that as part of a
move to use strong parameters [6] which was the same PR in which this
NOTE was added [7]. While that paragraph gives some useful context, I
don't think it's important/current enough to warrant continuing to live
as a comment in the code. Hopefully this commit note can serve a similar
purpose for future developers!

Regarding the fall back to the "normal" role, I note that the first two
tests in `InvitingUsersTest` [12] fail if this fall back is removed, so
the tests serve as documentation that can't get out-of-sync with the
code.

* 2nd paragraph, 2nd sentence

> In fact, a user accepting an invitation only needs to modify
`password` and `password_confirmation` so we could only permit those two
params for the `edit` and `update` actions.

At the time the NOTE was added [7] the `current_user_role` was used in
`UserParameterSanitiser` to determine which attributes of a `User` the
current user was allowed to change. A normal user could edit the
following attributes: `uid`, `name`, `email`, `password` and
`password_confirmation` [8]. This is still the case [9] and so the
suggestion is still valid, i.e. a user probably ought not to be able to
change their `uid` or `name` when accepting an invitation which they
could currently do by crafting suitable HTTP requests.

I'm going to bear this in mind as I make changes to the
`InvitationsController` as part of this Trello card [10]. I note that
this PR comment [11] suggests that the functionality deciding which
parameters should be available for which roles would probably be better
if it lived in a Pundit policy. I tend to agree and I think that would
make it easier to restrict parameters on a per-controller-action basis
too.

* Summary

All in all, given all the above, I'm happy to delete the comment!

[1]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L4
[2]: https://github.com/heartcombo/devise/blob/8b0b849a67c46b10827743aa0ccb0679d69e5396/app/controllers/devise_controller.rb#L99-L119
[3]: bffb80b9b2efbd9d428110898f682339136a9fa3
[4]: #352
[5]: 1bb38e533aa5d90fc8af769ac212f21477729cdb
[6]: #336
[7]: 3e939db4eff7ac551e835aafb40217e803b5be10
[8]: https://github.com/alphagov/signon/blob/3e939db4eff7ac551e835aafb40217e803b5be10/lib/roles/normal.rb#L4-L6
[9]: https://github.com/alphagov/signon/blob/591bfe2deca05804e5cba43221c4f8694e6d6be3/lib/roles/normal.rb#L3C1-L5
[10]: https://trello.com/c/umBDQZUc
[11]: https://github.com/alphagov/signon/pull/336#issuecomment-91215050
[12]: https://github.com/alphagov/signon/blob/591bfe2deca05804e5cba43221c4f8694e6d6be3/test/integration/inviting_users_test.rb#L6-L32