Update suspension reminder email content
Trello: https://trello.com/c/0oTag1bi

We've done some testing of this email and come up with this updated
content.

Notify, which we use to send emails, has limited markdown
support[1]. I've used the heading markup for "How to stop suspension",
and the `^` to call out the "please do not reply to this email".

[1] https://www.notifications.service.gov.uk/using-notify/formatting