Update suspension notification email content
Trello: https://trello.com/c/0oTag1bi

We've done some testing of this email and come up with this updated
content.

Notify, which we use to send emails, has limited markdown support[1],
notably it only supports two levels of heading. I've used the
second-level heading markup for "How to reactive your account" and "If
you're a managing editor or user manager". I've also used the the `^`
to call out the "please do not reply to this email".

[1] https://www.notifications.service.gov.uk/using-notify/formatting