Avoid calls to Object#try in InvitationsController
This is a small refactoring which I think is slightly easier to read.

A `User` has its role set to "normal" by default due to a default value
on the `users.role` database column. I did contemplate explicitly
setting the `User#role` to "normal", but I thought that would make
things harder if we ever wanted to change the default role.

I'm confident in making this change, because I know that two tests in
`InvitingUsersTest` [1] fail if the fallback is not in place.

[1]: https://github.com/alphagov/signon/blob/591bfe2deca05804e5cba43221c4f8694e6d6be3/test/integration/inviting_users_test.rb#L6-L32