Make InvitationsController#edit, #update & #destroy more explicit
Previously these actions were only implemented in the superclass and it
was far from obvious that this was the case. Although adding these
methods isn't strictly necessary, I think doing so makes things a bit
clearer.

I've also added a comment to make it clearer that the `edit` action
renders the custom `app/views/devise/invitations/edit.html.erb`
template.

Note that I did contemplate adding some test coverage for these actions
in InvitationsControllerTest, but since we're not customizing their
behaviour, we can probably safely rely on the tests within
devise_invitable. Also our InvitingUsersTest integration test provides
some good coverage of `edit` & `update` actions.

Note We don't seem to be using the `destroy` action anywhere in our
code. I did wonder if it was intended to be requested from a link in an
email, especially since it responds to GET requests, but it doesn't
appear in our `app/views/user_mailer/invitation_instructions.text.erb`
or in either of the equivalent templates in devise_invitable [1]; it
only appears in tests in devise_invitable. So perhaps we can remove it!

[1]: https://github.com/scambra/devise_invitable/tree/v2.0.8/app/views/devise/mailer