Make InvitationsController more like its superclass
The superclass, `Devise::InvitationsController`, prepends the
`authenticate_inviter!` before action [1] for the `new` & `create`
actions. This before action calls `authenticate_user!` anyway [2],
albeit with the `force` option set to `true`.

Although the `authenticate_user!` before action looked like it applied
to all actions, it was being bypassed by the superclass prepending the
`require_no_authentication` before action [3] for the `edit`, `update` &
`destroy` actions.

We've previously added the `resend` action which effectively does the
same as `create` but for an existing user who has already been invited.
So this needs the same protection as `new` & `create`. There's some
question in my mind about whether instead of using this action, we could
re-use the `create` action but with the `resend_invitation`
configuration option [4]. However, I'm not going to tackle that yet.

So, all in all, I think it's clearer and more consistent with the
superclass implementation to remove the `authenticate_user!` before
action from `InvitationsController` and add `authenticate_inviter!` as a
before action but for for `new`, `create` AND `resend` actions instead
of just for `new` & `create` as it is in the superclass.

I've also changed the `verify_authorized` after action from using
the `except` option to the `only` option to match the
`authenticate_inviter!` before action. While I accept this introduces a
risk that we'll forget to add any new action to these `only` option
arrays, I think the improved consistency and clarity is worth it.

[1]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L2
[2]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/lib/devise_invitable/controllers/helpers.rb#L21-L23
[3]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L4
[4]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/lib/devise_invitable.rb#L52-L59