Remove out-of-date comment from InvitationsController#create
This comment was out-of-date, because it talked about
`accepts_nested_attributes_for` which hasn't been used for some time.

I don't think the error it talks about is still pertinent either,
because the permissions are now created via
`InvitationsController#grant_default_permissions` which iterates through
the default supported permissions calling `User#grant_permission` for
each. The latter is idempotent, so if the user already has the
permission, it won't be added again. I suppose one minor issue is that
it won't remove any permissions that have become non-default permissions
since they were added to the user the first time round, but that doesn't
seem like a big issue.

Anyway, I'm happy to remove this somewhat confusing comment, because we
have a test in place [1] for the relevant behaviour. We can think about
whether we still need that behaviour separately.

[1]: https://github.com/alphagov/signon/blob/591bfe2deca05804e5cba43221c4f8694e6d6be3/test/controllers/invitations_controller_test.rb#L43-L50