Use `redirect_to` instead of `redirect_back_or_to`
We were only doing this for one of the actions in this controller, which
was resulting in an inconsistent experience for users.

Since there are soon going to be multiple routes into the "Change your
email or password" page (i.e. the page that uses this endpoint), it
might make sense to redirect back to the page that the user visited
*before* that one.  That's not what `redirect_back_or_to` was doing in
this case, though, it was redirecting back to the "Change your email or
password" page itself. Either way, it should be consistent.