Rename "user" to "inviter" in InvitationsControllerTest
I was finding reading the InvitationsControllerTest very confusing,
because there are two users involved in each of the tests - the inviter
and the invitee, but both were often referred to using text or variables
named "user". I've changed the relevant instances of "user" to "inviter"
and I plan to make a similar change to rename the remaining instances of
"user" to "invitee" in a subsequent commit.