Extract conditional page title/name to helper
I'm planning to make the 2SV setup page an Account sub-page. Without any
better ideas, I'd like the wording of the link from the Account page to
be consistent with the page's own existing title and heading.

The wording of the link on the Dashboard and the (title and heading of
the) 2SV page weren't exactly identical before this change, but the
difference seemed arbitrary enough not to worry about.