Build a temporary Manage permissions page
This is going to be linked to from the new Account page (for users who
are allowed to changing their own permissions).

There's an entire new permissions management area being developed as we
speak, so it doesn't make sense to put a lot of additional effort into
this version of it.

As a working placeholder, I've copied the permissions management section
of `Users#edit` and I've tried to keep the changes to a minimum.

Some things just didn't make sense in the new (current_user's account)
context. For example, I've tried to remove confusing references to the
user that's having their permissions changed -- in this case that's just
the current_user, changing their own permissions.