Point redirects to Account page, not to Dashboard
Instead of redirecting to the Dashboard, Account page-related actions
now redirect to the Account page on success.

The page that these actions are performed from can be accessed from
either the Dashboard or the Account page, but:

1. (I think) we can't tell which route the user took to get here
2. the Account page is now their canonical home

I can't say I really know where these actions would be best redirecting
back to in terms of the user experience. How they fit with a typical
user's expectations might depend on the specific action they're
performing, too.