Extract more contexts in InvitationsControllerTest
This introduces new contexts for when the invitee's organisation does or
does not require 2SV.

My main motivation for doing this was to be able to expand the test
names to explain that the "2SV form" (i.e. the one rendered by
`UsersController#require_2sv`) only allows the inviter to choose whether
to require 2SV when the invitee's organisation does not require 2SV.
Previously I don't think this was very obvious - or at least I found the
tests rather confusing.

Two of the new contexts have the same name where the organisation does
not require 2SV. I thought I'd do that in a subsequent commit to make
the diffs less confusing.