Make InvitationsControllerTest test names more explanatory
This improves some test name so that they explain why the inviter cannot
choose whether 2SV is required for any kind of admin invitee even if
they will be assigned to an organisation for which 2SV is not required,
i.e. because 2SV is required for all kinds of admin invitees
irrespective of the organanisation-level setting.