Inline form_fields partial into new invitation template
This inlines `app/views/users/_form_fields.html.erb` into
`app/views/devise/invitations/new.html.erb`, but leaves the former in
place so it can still be used by `UsersController#edit`.

There's a bunch of code in this partial which is irrelevant for a new
user. I plan to remove that from
`app/views/devise/invitations/new.html.erb` in subsequent commits.
Similarly there is almost certainly some code in the partial which is
irrelevant for a persisted user. I'll try to remove that too.

While this comes at the cost of some duplication, I think it'll be worth
it to make it clearer what user attributes can be set or updated in each
of the scenarios.