Simplify logic in new invitation template & user form fields partial
Previously both the `if` *and* the `elsif` included the
`policy(User).assign_role?` conditional. By moving the rest of the `if`
condition into its own nested `if` and pairing it with an `else` we can
make things clearer.

Ideally I would've done this *before* I inlined the form fields into the
new invitation template, but it would be awkward to rewrite the git
history, so I'm taking the lazy option and updating it in both places.