Add test coverage for InvitationsController#update
It's not immediately obvious but this is the action which is used for
accepting an invitation at which point the user is required to set their
password.

While we're just using the inherited
`Devise::InvitationsController#update` [1] unchanged, we *are*
customizing some of the methods that it uses, e.g. `#resource_params` &
`#update_resource_params`. So while to some extent we can rely on the
tests in `devise_invitable`, I want to get some more confidence about
the current behaviour.

In particular I find the way the code decides whether certain params are
permitted (for bulk assignment) or not is quite hard to follow, because
it's spread across a number of methods & classes, e.g.
* `InvitationsController#resource_params`
* `InvitationsController#unsanitised_user_params`
* `InvitationsController#current_user_role`
* `UserParameterSanitiser`
* `Roles::***.permitted_user_params`

For example, the reason the invitation acceptance form can't be hacked
to update the invitee's role (which seems quite important) is because
`InvitationsController#current_user_role` will be "normal" and so
`UserParameterSanitiser#sanitise` will exclude the `role` param, because
`Roles::Normal.permitted_user_params` does not include `role`. This
isn't immediately obvious!

[1]: https://github.com/scambra/devise_invitable/blob/v2.0.8/app/controllers/devise/invitations_controller.rb