Prevent users editing their details on users/edit pages
Users should edit their own details using their account page
functionality. This commit ensures that no one can continue using the
/users pages to edit their own account.

In the UserPolicy test for superadmins and admins, I have to skip the
`create` action because these roles should continue to be able to create
new users.

This includes a fix for a bug in the _form_fields partial. The partial
was previously checking whether the signed in user can mandate_2sv for
themselves instead of checking whether the signed in user can
mandate_2sv for the user they're editing. The change to UserPolicy in
this commit highlighted the bug because users can no longer edit their
details (including mandate_2sv) through the users controller.

I've had to update two tests in users_controller_test so that we're
mimicking a user editing another user, instead of the user editing
themselves.