Use built-in strong params for InvitationsController#update
It's not immediately obvious but this is the action which is used for
accepting an invitation at which point the user is required to set their
password.

Previously we were overriding
`Devise::InvitationsController#update_resource_params` to return the
params from `InvitationsController#resource_params` which in turn was
overriding `Devise::InvitationsController#resource_params`.
`InvitationsController#resource_params` was using
`UserParameterSanitiser#sanitise` to further restrict the params
permitted in `InvitationsController#unsanitised_user_params`.

The latter permitted the following keys:
* `name`
* `email`
* `organisation_id`
* `invitation_token`
* `password`
* `password_confirmation`
* `require_2sv`
* `role`
* `supported_permission_ids`

And then since `current_user` is always `nil` in `InvitationsController#update`,
`UserParameterSanitiser#sanitise` was further restricting the permitted
params by only allowing keys returned from `Roles::Normal.permitted_user_params`:
* `uid`
* `name`
* `email`
* `password`
* `password_confirmation`

As you can see `uid` is never permitted by
`InvitationsController#unsanitised_user_params` and so the only keys
permitted by `UserParameterSanitiser#sanitise` were:
* `name`
* `email`
* `password`
* `password_confirmation`

`InvitationsController#resource_params` was then adding
`invitation_token`, because it was handling the `update` action. Giving
a final list of:
* `name`
* `email`
* `password`
* `password_confirmation`
* `invitation_token`

This seems like an extremely complicated and overly permissive way to
permit the params that are actually needed for this action, i.e.
`invitation_token`, `password` & `password_confirmation` which are the
fields in the form in `app/views/devise/invitations/edit.html.erb` and
thence `app/views/devise/passwords/_change_password_panel.html.erb`
(`current_password` is not rendered as part of
`InvitationsController#edit`, because the `updating_password` local is
not set).

In this commit I've attempted to come up with a simpler solution which
is also more in keeping with `devise_invitable` as follows:

* I've renamed `InvitationsController#resource_params` to
`#invite_params` and change the implementation of `#create` so it uses
the renamed method. The new method name matches the method that's used
in the superclass [1]. The idea is that this method is only used to
provide params to the `#create` action and not the `#update` action.

* I've removed the custom implementation of `#update_resource_params`
and instead fallback to the implementation in the superclass [2]. This
uses the Strong Parameters support built into `devise_invitable` [3] to
permit `invitation_token`, `password` & `password_confirmation`, i.e.
the only params that we require.

* Since the method renamed to `#invite_params` is no longer used for the
`#update` action, we can remove the `if`/`else` statement. And since
the `#invitation_token` method is not longer used in that statement, it
can also be removed.

These changes have the effect of tightening up the list of permitted
params so that `name` & `email` are no longer permitted. Previously it
was theoretically possible to hack the invitation acceptance form to
change your name & email, but I strongly suspect this was not
intentional and possibly even a security risk.

As an aside, it was rather odd that the fallback value of
`#invitation_token` was `{}`! I would've epxected it to be "" or just
`nil`.

[1]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L106C9-L108
[2]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L110-L112
[3]: https://github.com/scambra/devise_invitable/blob/v2.0.8/README.rdoc#label-Strong+Parameters