Remove redundant conditional in new invitation template
This template is only rendered from `InvitationsController#new` and the
`authorize User` statement calls `UserPolicy#new?` which only allows
access to admins and superadmins.

The conditional in the template was calling
`UserPolicy#assign_organisations?` which is aliased to
`UserPolicy#new?`, i.e. anyone who can view the page is allowed to
assign an organisation to the invitee.

Thus the conditional is redundant and can be removed.