Prevent duplicate rows in /users when filters applied
Trello: https://trello.com/c/i27MQaus

This fixes an issue with duplication of a user in the table on
/users (and the corresponding CSV export) when multiple permissions
for the same application were selected in the filter.

For example if a user had the "GDS editor" and "signon" permission for
the "Transition" app and we selected both of those permissions in the
filter drop down, this user would appear twice in table.

The solution is to call `distinct` on the `User#with_permission`
scope, which is in turn called by `UsersFilter#users`. I've decided to
add a test case in both places to avoid a regression in `UsersFilter`
in case it changes in the future.

Note we have to add the `unscoped` method to `SupportedPermission` in
the `User#with_permission` scope since `SupportedPermission` has an
`order(:name)` `default_scope` and calling `distinct` on an order
query is not supported by MySQL:

    Mysql2::Error: Expression #1 of ORDER BY clause is not in SELECT
    list, references column 'signon_test.supported_permissions.name' which
    is not in SELECT list; this is incompatible with DISTINCT