Use built-in strong params for InvitationsController#create
It's not immediately obvious but this is the action which is used for
creating a user and sending them an invitation.

Previously we were overriding
`Devise::InvitationsController#invite_params` to use
`UserParameterSanitiser#sanitise` to further restrict the params
permitted in `InvitationsController#unsanitised_user_params`.

The latter permitted the following keys:
* `name`
* `email`
* `organisation_id`
* `invitation_token`
* `password`
* `password_confirmation`
* `require_2sv`
* `role`
* `supported_permission_ids`

And then since `current_user` is always and admin or a superadmin in
`InvitationsController#create`, `UserParameterSanitiser#sanitise` was
further restricting the permitted params by only allowing keys returned
from `Roles::Admin.permitted_user_params` or
`Roles::Superadmin.permitted_user_params`:
* `uid`
* `name`
* `email`
* `organisation_id`
* `password`
* `password_confirmation`
* `require_2sv`
* `role` (only for superadmins)
* `supported_permission_ids`
* `unconfirmed_email`
* `confirmation_token`

As you can see `uid`, `invitation_token`, `unconfirmed_email` &
`confirmation_token` were never permitted by
`InvitationsController#unsanitised_user_params` and so the only keys
permitted by `UserParameterSanitiser#sanitise` were:
* `name`
* `email`
* `organisation_id`
* `password`
* `password_confirmation`
* `require_2sv`
* `role` (only for superadmins)
* `supported_permission_ids`

This seems like an extremely complicated and overly permissive way to
permit the params that are actually needed for this action, i.e. `name`,
`email`, `organisation_id`, `supported_permission_ids` & `role` (the
latter only for superadmins) which are the fields in the form in
`app/views/devise/invitations/new.html.erb` and thence
`app/views/shared/_user_permissions.html.erb`.

In this commit I've attempted to come up with a simpler solution which
is also more in keeping with `devise_invitable` as follows:

* I've removed `InvitationsController#invite_params` which means that
  the `#create` action now relies on the superclass implementation [1].
  This only permits the `email` param by default.

* I've added a new `configure_permitted_parameters` before action (just
  for the `create` action) as recommended in the `devise_invitable`
  documentation for Strong Parameters [2]. The
  `#configure_permitted_parameters` method permits `name`,
  `organisation_id` & `supported_permission_ids`. It also permits `role`,
  but only if the current user is allowed to assign a role, i.e. they are
  a superadmin (this is the only difference between the keys returned by
  `Roles::Superadmin.permitted_user_params` &
  `Roles::Admin.permitted_user_params`). I believe these are the only
  params that should be submitted by the
  `app/views/devise/invitations/new.html.erb` form (see above).

* I've removed `#unsanitised_user_params` & `#current_user_role` which
  are both now redundant.

These changes have the effect of tightening up the list of permitted
params so that `password`, `password_confirmation` & `require_2sv` are
no longer permitted. Previously it was theoretically possible to hack
the new invitation form to change your these attributes on the created
user, but I strongly suspect this was not intentional.

[1]: https://github.com/scambra/devise_invitable/blob/db2f9a1c0e87addb196d1b82920370f4230c1157/app/controllers/devise/invitations_controller.rb#L106C9-L108
[2]: https://github.com/scambra/devise_invitable/blob/v2.0.8/README.rdoc#label-Strong+Parameters