Simplify authorization for InvitationsController#create
Previously we were only temporarily instantiating a user in order to
(implicitly) pass it (as `record`) to `UserPolicy#create?`. The
implementation of the latter was the same as `UserPolicy#new?` except
that it prevented admins from creating a superadmin invitee.

However, since only superadmins (and not admins) are allowed to assign a
role (see `UserPolicy#assign_role?`, admins can only ever create
invitees with the "normal" role.

So it seems simpler to change the authorization for the `#create` action
to work exactly the same way as the `#new` action. It also has the
benefit that we can legitimately remove the slightly confusiing comment
about a "workaround".

I'm not 100% sure the changes to `UserPolicyTest` are correct, because I
found them rather confusing. However, I've tried to change the tests for
`UserPolicy#create?` to be the same as for `UserPolicy#new?`. To do this
I've moved `create` from `user_management_actions` to
`primary_management_actions` which is where `new` lives. And I've had to
remove two existing explicit tests for `UserPolicy#create?`, because
`primary_management_actions` was creating duplicate tests.