Remove call to symbolize_keys from InvitationsController#create
I suspect this was previously necessary (if it ever was), because we
were converting an instance of `ActionController::Parameters` to a
`Hash` by calling `#to_h` in what was the `#resource_params` before the
set of changes in this PR.

However, now that we're using the `#invite_params` implemented in the
superclass, we're now working directly with an instance of
`ActionController::Parameters` which seems to allow access via either
string or symbol keys. Thus the call to `#symbolize_keys` is redundant
and can be removed.