Keep selected permissions on validation error in new invitation
This fixes a bug where any non-signin permissions that were selected
before the form was submitted were being "lost" when a validation error
occurred and the form was re-rendered.

The problem was that `User#permission_ids_for` only worked for users
that had already been persisted. I've fixed it by changing the
implementation to use the `User#application_permissions` in memory.

Although `User#permission_ids_for` is used on a number of pages via the
`app/views/shared/_user_permissions.html.erb` partial, I'm pretty sure
all the other places related to persisted users. So I only added a
failing test in InvitationsControllerTest which forced me to fix the
problem.

Note that signin permissions were not affected, because they are
rendered in a checkbox separate from the select. However, I've added a
test for this for completeness. Interestingly the reason this wasn't
broken was that `User#has_access_to?` works OK even if the user is not
persisted, because it operates on `User#application_permissions` in a
similar way to my new implementation of `User#permission_ids_for`.

It's possible that this change will lead to page that use the
`app/views/shared/_user_permissions.html.erb` partial being a bit slower
to load. However, it doesn't seem to be a significant problem when I run
the app locally. If it does turn out to be a problem, we could add an if
condition in `User#permission_ids_for` which checks whether the user is
persisted or not.

I've also taken the opportunity to add some test coverage for retaining
other form field values when a validation error occurs.